syntest path traveral - possible fix

+++ ProEstimator.Admin/Controllers/DueInvoiceReportController.cs
@@ -2,6 +2,7 @@
 using Kendo.Mvc.UI;
 using ProEstimator.Admin.ViewModel.DueInvoiceReport;
 using ProEstimator.Business.Logic;
+using ProEstimator.Business.Model;
 using ProEstimatorData;
 using ProEstimatorData.DataModel;
 using ProEstimatorData.DataModel.Admin;
@@ -10,6 +11,7 @@ using System.Collections.Generic;
 using System.Configuration;
 using System.Data;
 using System.IO;
+using System.Text.RegularExpressions;
 using System.Web.Mvc;
 
 namespace ProEstimator.Admin.Controllers
@@ -60,6 +62,18 @@ namespace ProEstimator.Admin.Controllers
                 rangeStartFilter = string.IsNullOrEmpty(rangeStartFilter) ? null : rangeStartFilter;
                 rangeEndFilter = string.IsNullOrEmpty(rangeEndFilter) ? null : rangeEndFilter;
 
+                if (rangeStartFilter.Contains("/") || rangeStartFilter.Contains(@"\") || rangeStartFilter.Contains("$") ||
+                    rangeStartFilter.Contains("..") || rangeStartFilter.Contains("?"))
+                {
+                    return null;
+                }
+
+                if (rangeEndFilter.Contains("/") || rangeEndFilter.Contains(@"\") || rangeEndFilter.Contains("$") ||
+                    rangeEndFilter.Contains("..") || rangeEndFilter.Contains("?"))
+                {
+                    return null;
+                }
+
                 autoPayFilter = string.Compare(autoPayFilter, "false", true) == 0 ? null : autoPayFilter;
                 paidFilter = string.Compare(paidFilter, "false", true) == 0 ? null : paidFilter;
                 hasStripeInfoFilter = string.Compare(hasStripeInfoFilter, "false", true) == 0 ? null : hasStripeInfoFilter;
@@ -81,10 +95,30 @@ namespace ProEstimator.Admin.Controllers
                     Directory.CreateDirectory(adminFolder);
                 }
 
-                if (System.IO.File.Exists(diskPath))
+                Regex regex = new Regex(@"([a-zA-Z0-9\s_\\.\-:])+(.dat)$");
+                Match match = regex.Match(diskPath);
+
+                if (match.Success)
                 {
-                    System.IO.File.Delete(diskPath);
+                    if (System.IO.File.Exists(diskPath) && Path.GetFullPath(diskPath).StartsWith(@"C:\ProUserContent\Admin", StringComparison.OrdinalIgnoreCase))
+                    {
+                        if (System.IO.File.Exists(diskPath))
+                        {
+                            System.IO.File.Delete(diskPath);
+                        }
+                    }
+                    else
+                    {
+                        //return "File not found";
+                        return null;
+                    }
                 }
+                else
+                {
+                    //return "File name not valid";
+                    return null;
+                }
+
 
                 spreadsheetWriter.WriteSpreadshet(table, diskPath);
